console: Refactor connect in hw version select
