+++ pkg/webui/locales/en.json
@@ -360,9 +360,6 @@
   "console.components.uplink-form.index.simulateUplink": "Simulate uplink",
   "console.components.uplink-form.index.payloadDescription": "The desired payload bytes of the uplink message",
   "console.components.uplink-form.index.uplinkSuccess": "Uplink sent",
-  "console.components.uplink-form.uplink-form.simulateUplink": "Simulate uplink",
-  "console.components.uplink-form.uplink-form.payloadDescription": "The desired payload bytes of the uplink message",
-  "console.components.uplink-form.uplink-form.uplinkSuccess": "Uplink sent",
   "console.components.webhook-form.index.idPlaceholder": "my-new-webhook",
   "console.components.webhook-form.index.messageInfo": "For each enabled event type an optional path can be defined which will be appended to the base URL",
   "console.components.webhook-form.index.deleteWebhook": "Delete Webhook",
@@ -500,9 +497,8 @@
   "console.containers.device-profile-section.device-selection.band-select.index.title": "Profile (Region)",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.title": "End device brand",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.noOptionsMessage": "No matching brand found",
-  "console.containers.device-profile-section.device-selection.fw-version-select.fw-version-select.title": "Firmware Ver.",
   "console.containers.device-profile-section.device-selection.fw-version-select.index.title": "Firmware Ver.",
-  "console.containers.device-profile-section.device-selection.hw-version-select.hw-version-select.title": "Hardware Ver.",
+  "console.containers.device-profile-section.device-selection.hw-version-select.index.title": "Hardware Ver.",
   "console.containers.device-profile-section.device-selection.model-select.model-select.noOptionsMessage": "No matching model found",
   "console.containers.device-profile-section.hints.other-hint.hintTitle": "Your end device will be added soon!",
   "console.containers.device-profile-section.hints.other-hint.hintMessage": "We're sorry, but your device is not yet part of The LoRaWAN Device Repository. You can use <b>enter end device specifics manually</b> option above, using the information your end device manufacturer provided e.g. in the product's data sheet. Please also refer to our documentation on <GuideLink>Adding Devices</GuideLink>.",
@@ -510,9 +506,6 @@
   "console.containers.device-profile-section.hints.progress-hint.hintNoSupportMessage": "Cannot find your exact end device? Try <b>enter end device specifics manually</b> option above.",
   "console.containers.device-template-format-select.index.title": "File format",
   "console.containers.device-template-format-select.index.warning": "End device template formats unavailable",
-  "console.containers.device-title-section.device-title-section.uplinkDownlinkTooltip": "The number of sent uplinks and received downlinks of this end device since the last frame counter reset.",
-  "console.containers.device-title-section.device-title-section.lastSeenAvailableTooltip": "The elapsed time since the network registered the last activity of this end device. This is determined from sent uplinks, confirmed downlinks or (re)join requests.{lineBreak}The last activity was received at {lastActivityInfo}",
-  "console.containers.device-title-section.device-title-section.noActivityTooltip": "The network has not registered any activity from this end device yet. This could mean that your end device has not sent any messages yet or only messages that cannot be handled by the network, e.g. due to a mismatch of EUIs or frequencies.",
   "console.containers.device-title-section.index.uplinkDownlinkTooltip": "The number of sent uplinks and received downlinks of this end device since the last frame counter reset.",
   "console.containers.device-title-section.index.lastSeenAvailableTooltip": "The elapsed time since the network registered the last activity of this end device. This is determined from sent uplinks, confirmed downlinks or (re)join requests.{lineBreak}The last activity was received at {lastActivityInfo}",
   "console.containers.device-title-section.index.noActivityTooltip": "The network has not registered any activity from this end device yet. This could mean that your end device has not sent any messages yet or only messages that cannot be handled by the network, e.g. due to a mismatch of EUIs or frequencies.",
@@ -538,17 +531,6 @@
   "console.containers.gateway-location-form.gateway-location-form.locationFromStatusMessage": "Location set automatically from status messages",
   "console.containers.gateway-location-form.gateway-location-form.setLocationManually": "Set location manually",
   "console.containers.gateway-location-form.gateway-location-form.noLocationSetInfo": "This gateway has no location information set",
-  "console.containers.gateway-location-form.index.updateLocationFromStatus": "Update from status messages",
-  "console.containers.gateway-location-form.index.updateLocationFromStatusDescription": "Update the location of this gateway based on incoming status messages",
-  "console.containers.gateway-location-form.index.setGatewayLocation": "Gateway antenna location settings",
-  "console.containers.gateway-location-form.index.locationSource": "Location source",
-  "console.containers.gateway-location-form.index.locationPrivacy": "Location privacy",
-  "console.containers.gateway-location-form.index.placement": "Placement",
-  "console.containers.gateway-location-form.index.indoor": "Indoor",
-  "console.containers.gateway-location-form.index.outdoor": "Outdoor",
-  "console.containers.gateway-location-form.index.locationFromStatusMessage": "Location set automatically from status messages",
-  "console.containers.gateway-location-form.index.setLocationManually": "Set location manually",
-  "console.containers.gateway-location-form.index.noLocationSetInfo": "This gateway has no location information set",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.index.requireAuthenticatedConnectionDescription": "Select which information can be seen by other network participants, including {packetBrokerURL}",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.index.shareGatewayInfoDescription": "Choose this option eg. if your gateway is powered by {loraBasicStationURL}",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.validation-schema.validateEntry": "There must be at least one selected frequency plan ID.",
@@ -600,8 +582,6 @@
   "console.containers.organizations-table.index.purgeFail": "There was an error and the organization could not be purged",
   "console.containers.owners-select.index.title": "Owner",
   "console.containers.owners-select.index.warning": "There was an error and the list of organizations could not be displayed",
-  "console.containers.owners-select.owners-select.title": "Owner",
-  "console.containers.owners-select.owners-select.warning": "There was an error and the list of organizations could not be displayed",
   "console.containers.packet-broker-networks-table.index.nonDefaultPolicies": "Networks with non-default policies",
   "console.containers.packet-broker-networks-table.index.search": "Search by tenant ID or name",
   "console.containers.packet-broker-networks-table.index.forwarderPolicy": "Their routing policy towards us",
