+++ pkg/webui/locales/ja.json
@@ -357,12 +357,9 @@
   "console.components.routing-policy-form.index.saveDefaultPolicy": "",
   "console.components.routing-policy-form.index.useSpecificPolicy": "ネットワーク固有のルーティングポリシーを使用",
   "console.components.routing-policy-form.index.doNotUseAPolicy": "このネットワークには、ルーティングポリシーを使用しないでください",
-  "console.components.uplink-form.index.simulateUplink": "",
-  "console.components.uplink-form.index.payloadDescription": "",
-  "console.components.uplink-form.index.uplinkSuccess": "",
-  "console.components.uplink-form.uplink-form.simulateUplink": "アップリンクのシミュレーション",
-  "console.components.uplink-form.uplink-form.payloadDescription": "アップリンクメッセージの希望するペイロードバイト数",
-  "console.components.uplink-form.uplink-form.uplinkSuccess": "アップリンク送信済",
+  "console.components.uplink-form.index.simulateUplink": "アップリンクのシミュレーション",
+  "console.components.uplink-form.index.payloadDescription": "アップリンクメッセージの希望するペイロードバイト数",
+  "console.components.uplink-form.index.uplinkSuccess": "アップリンク送信済",
   "console.components.webhook-form.index.idPlaceholder": "my-new-webhook",
   "console.components.webhook-form.index.messageInfo": "有効なメッセージタイプごとに、オプションのパスをベースURLに合わせて定義することができます",
   "console.components.webhook-form.index.deleteWebhook": "Webhookの削除",
@@ -500,9 +497,8 @@
   "console.containers.device-profile-section.device-selection.band-select.index.title": "プロフィール（リージョン）",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.title": "",
   "console.containers.device-profile-section.device-selection.brand-select.brand-select.noOptionsMessage": "",
-  "console.containers.device-profile-section.device-selection.fw-version-select.fw-version-select.title": "",
   "console.containers.device-profile-section.device-selection.fw-version-select.index.title": "",
-  "console.containers.device-profile-section.device-selection.hw-version-select.hw-version-select.title": "",
+  "console.containers.device-profile-section.device-selection.hw-version-select.index.title": "",
   "console.containers.device-profile-section.device-selection.model-select.model-select.noOptionsMessage": "",
   "console.containers.device-profile-section.hints.other-hint.hintTitle": "お客様のエンドデバイスはすぐに追加されます！",
   "console.containers.device-profile-section.hints.other-hint.hintMessage": "申し訳ありませんが、あなたのデバイスはまだLoRaWANデバイスリポジトリの一部ではありません。エンドデバイスの製造元が提供する情報（製品のデータシートなど）を使用して、上記のenter end device specifics manuallyオプションを使用することができます。また、デバイスの追加に関するドキュメントも参照してください",
@@ -510,11 +506,8 @@
   "console.containers.device-profile-section.hints.progress-hint.hintNoSupportMessage": "",
   "console.containers.device-template-format-select.index.title": "ファイルフォーマット",
   "console.containers.device-template-format-select.index.warning": "エンドデバイスのテンプレートフォーマットが利用できません",
-  "console.containers.device-title-section.device-title-section.uplinkDownlinkTooltip": "前回のフレームカウンタリセット以降、このエンドデバイスの送信アップリンクと受信ダウンリンクの数です",
-  "console.containers.device-title-section.device-title-section.lastSeenAvailableTooltip": "ネットワークがこのエンドデバイスの最後のアクティビティを登録してから経過した時間です。これは、送信されたアップリンク、確認されたダウンリンク、または（再）参加要求から判断されます。{lineBreak}最後のアクティビティは{lastActivityInfo}で受信します",
-  "console.containers.device-title-section.device-title-section.noActivityTooltip": "ネットワークは、このエンドデバイスからのアクティビティをまだ登録していません。これは、エンドデバイスがまだメッセージを送信していないか、EUIや周波数の不一致など、ネットワークで処理できないメッセージしか送信していないことを意味する可能性があります",
-  "console.containers.device-title-section.index.uplinkDownlinkTooltip": "",
-  "console.containers.device-title-section.index.lastSeenAvailableTooltip": "",
+  "console.containers.device-title-section.index.uplinkDownlinkTooltip": "前回のフレームカウンタリセット以降、このエンドデバイスの送信アップリンクと受信ダウンリンクの数です",
+  "console.containers.device-title-section.index.lastSeenAvailableTooltip": "ネットワークがこのゲートウェイの最後のアクティビティを登録してから経過した時間です。これは、このゲートウェイの受信したアップリンク、または送信したステータスメッセージから決定されます",
   "console.containers.device-title-section.index.noActivityTooltip": "",
   "console.containers.devices-table.index.otherClusterTooltip": "このエンドデバイスは、別のクラスタ（`{host}`）に登録されています。このデバイスにアクセスするには、このエンドデバイスが登録されているクラスタのコンソールを使用します",
   "console.containers.freq-plans-select.utils.warning": "",
@@ -522,7 +515,7 @@
   "console.containers.freq-plans-select.utils.selectFrequencyPlan": "",
   "console.containers.freq-plans-select.utils.addFrequencyPlan": "",
   "console.containers.freq-plans-select.utils.frequencyPlanDescription": "",
-  "console.containers.gateway-connection.gateway-connection.lastSeenAvailableTooltip": "ネットワークがこのゲートウェイの最後のアクティビティを登録してから経過した時間です。これは、このゲートウェイの受信したアップリンク、または送信したステータスメッセージから決定されます",
+  "console.containers.gateway-connection.gateway-connection.lastSeenAvailableTooltip": "",
   "console.containers.gateway-connection.gateway-connection.disconnectedTooltip": "ゲートウェイは現在、ゲートウェイサーバーとの TCP 接続を確立していません。まれに）UDPベースのゲートウェイの場合、これはゲートウェイが過去30秒以内にpull/pushデータ要求を開始しなかったことを意味することもあります",
   "console.containers.gateway-connection.gateway-connection.connectedTooltip": "このゲートウェイはゲートウェイサーバーに接続されていますが、ネットワークはまだゲートウェイからのアクティビティ（アップリンクやステータスメッセージの送信）を登録していません",
   "console.containers.gateway-connection.gateway-connection.otherClusterTooltip": "このゲートウェイは、このクラスタのメッセージを処理しない外部のゲートウェイサーバーに接続されています。そのため、このゲートウェイからのアクティビティを見ることはできません",
@@ -538,17 +531,6 @@
   "console.containers.gateway-location-form.gateway-location-form.locationFromStatusMessage": "",
   "console.containers.gateway-location-form.gateway-location-form.setLocationManually": "",
   "console.containers.gateway-location-form.gateway-location-form.noLocationSetInfo": "",
-  "console.containers.gateway-location-form.index.updateLocationFromStatus": "",
-  "console.containers.gateway-location-form.index.updateLocationFromStatusDescription": "",
-  "console.containers.gateway-location-form.index.setGatewayLocation": "",
-  "console.containers.gateway-location-form.index.locationSource": "",
-  "console.containers.gateway-location-form.index.locationPrivacy": "",
-  "console.containers.gateway-location-form.index.placement": "",
-  "console.containers.gateway-location-form.index.indoor": "",
-  "console.containers.gateway-location-form.index.outdoor": "",
-  "console.containers.gateway-location-form.index.locationFromStatusMessage": "",
-  "console.containers.gateway-location-form.index.setLocationManually": "",
-  "console.containers.gateway-location-form.index.noLocationSetInfo": "",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.index.requireAuthenticatedConnectionDescription": "{packetBrokerURL}など、他のネットワーク参加者が見ることができる情報を選択します",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.index.shareGatewayInfoDescription": "ゲートウェイが{loraBasicStationURL}で駆動している場合など、このオプションを選択します",
   "console.containers.gateway-onboarding-form.gateway-provisioning-form.gateway-registration-form-section.validation-schema.validateEntry": "",
@@ -600,8 +582,6 @@
   "console.containers.organizations-table.index.purgeFail": "エラーが発生したため、組織をパージすることができませんでした",
   "console.containers.owners-select.index.title": "",
   "console.containers.owners-select.index.warning": "",
-  "console.containers.owners-select.owners-select.title": "",
-  "console.containers.owners-select.owners-select.warning": "",
   "console.containers.packet-broker-networks-table.index.nonDefaultPolicies": "デフォルトでないポリシーを持つネットワーク",
   "console.containers.packet-broker-networks-table.index.search": "テナントID、テナント名で検索",
   "console.containers.packet-broker-networks-table.index.forwarderPolicy": "私たちに対する彼らのルーティングポリシー",
